# api/openapi/things.yml
...
openapi: 3.0.3
info:
  title: Magistrala Things Service
  description: |
    This is the Things Server based on the OpenAPI 3.0 specification.  It is the HTTP API for managing platform things and channels. You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    Some useful links:
    - [The Magistrala repository](https://github.com/absmach/magistrala)
  contact:
    email: info@abstractmachines.fr
  license:
    name: Apache 2.0
    url: https://github.com/absmach/magistrala/blob/main/LICENSE
  version: 0.14.0
servers:
  - url: http://localhost:9000
  - url: https://localhost:9000
...
paths:
...
  /things/{thingID}:
    get:
      operationId: getThing
      summary: Retrieves thing info
      description: |
        Retrieves a specific thing that is identifier by the thing ID.
      tags:
        - Things
      parameters:
        - $ref: "#/components/parameters/ThingID"
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/ThingRes"
        "401":
          description: Missing or invalid access token provided.
        "403":
          description: Failed to perform authorization over the entity.
        "404":
          description: A non-existent entity request.
        "422":
          description: Database can't process request.
        "500":
          $ref: "#/components/responses/ServiceError"

    patch:
      operationId: updateThing
      summary: Updates name and metadata of the thing.
      description: |
        Update is performed by replacing the current resource data with values
        provided in a request payload. Note that the thing's type and ID
        cannot be changed.
      tags:
        - Things
      parameters:
        - $ref: "#/components/parameters/ThingID"
      requestBody:
        $ref: "#/components/requestBodies/ThingUpdateReq"
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/ThingRes"
        "400":
          description: Failed due to malformed JSON.
        "401":
          description: Missing or invalid access token provided.
        "403":
          description: Failed to perform authorization over the entity.
        "404":
          description: Failed due to non existing thing.
        "409":
          description: Failed due to using an existing identity.
        "415":
          description: Missing or invalid content type.
        "422":
          description: Database can't process request.
        "500":
          $ref: "#/components/responses/ServiceError"
    delete:
      summary: Delete thing for a thing with the given id.
      description: |
        Delete thing removes a thing with the given id from repo
        and removes all the policies related to this thing.
      tags:
        - Things
      parameters:
        - $ref: "#/components/parameters/ThingID"
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Thing deleted.
        "401":
          description: Missing or invalid access token provided.
        "403":
          description: Unauthorized access to thing id.
        "404":
          description: Missing thing.
        "500":
          $ref: "#/components/responses/ServiceError"
...
  /things/{thingID}/secret:
    patch:
      operationId: updateThingSecret
      summary: Updates Secret of the identified thing.
      description: |
        Updates secret of the identified in thing. Secret is updated using
        authorization token and the new received info. Update is performed by replacing current key with a new one.
      tags:
        - Things
      parameters:
        - $ref: "#/components/parameters/ThingID"
      requestBody:
        $ref: "#/components/requestBodies/ThingUpdateSecretReq"
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/ThingRes"
        "400":
          description: Failed due to malformed JSON.
        "401":
          description: Missing or invalid access token provided.
        "403":
          description: Failed to perform authorization over the entity.
        "404":
          description: Failed due to non existing thing.
        "409":
          description: Specified key already exists.
        "415":
          description: Missing or invalid content type.
        "422":
          description: Database can't process request.
        "500":
          $ref: "#/components/responses/ServiceError"
...
  /channels/{chanID}:
    get:
      operationId: getChannel
      summary: Retrieves channel info.
      description: |
        Gets info on a channel specified by id.
      tags:
        - Channels
      parameters:
        - $ref: "#/components/parameters/chanID"
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/ChannelRes"
        "400":
          description: Failed due to malformed channel's ID.
        "401":
          description: Missing or invalid access token provided.
        "403":
          description: Failed to perform authorization over the entity.
        "404":
          description: Channel does not exist.
        "422":
          description: Database can't process request.
        "500":
          $ref: "#/components/responses/ServiceError"

    put:
      operationId: updateChannel
      summary: Updates channel data.
      description: |
        Update is performed by replacing the current resource data with values
        provided in a request payload. Note that the channel's ID will not be
        affected.
      tags:
        - Channels
      parameters:
        - $ref: "#/components/parameters/chanID"
      security:
        - bearerAuth: []
      requestBody:
        $ref: "#/components/requestBodies/ChannelUpdateReq"
      responses:
        "200":
          $ref: "#/components/responses/ChannelRes"
        "400":
          description: Failed due to malformed query parameters.
        "401":
          description: Missing or invalid access token provided.
        "403":
          description: Failed to perform authorization over the entity.
        "404":
          description: Channel does not exist.
        "409":
          description: Failed due to using an existing identity.
        "415":
          description: Missing or invalid content type.
        "422":
          description: Database can't process request.
        "500":
          $ref: "#/components/responses/ServiceError"
    delete:
      summary: Delete channel for given channel id.
      description: |
        Delete channel remove given channel id from repo
        and removes all the policies related to channel.
      tags:
        - Channels
      parameters:
        - $ref: "#/components/parameters/chanID"
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Channel deleted.
        "401":
          description: Missing or invalid access token provided.
        "403":
          description: Unauthorized access to thing id.
        "404":
          descripttion: A non-existent entity request.
        "500":
          $ref: "#/components/responses/ServiceError"
...
components:
  schemas:
...
    Thing:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: bb7edb32-2eac-4aad-aebe-ed96fe073879
          description: Thing unique identifier.
        name:
          type: string
          example: thingName
          description: Thing name.
        tags:
          type: array
          minItems: 0
          items:
            type: string
          example: ["tag1", "tag2"]
          description: Thing tags.
        domain_id:
          type: string
          format: uuid
          example: bb7edb32-2eac-4aad-aebe-ed96fe073879
          description: ID of the domain to which thing belongs.
        credentials:
          type: object
          properties:
            identity:
              type: string
              example: thingidentity
              description: Thing Identity for example email address.
            secret:
              type: string
              example: bb7edb32-2eac-4aad-aebe-ed96fe073879
              description: Thing secret password.
        metadata:
          type: object
          example: { "model": "example" }
          description: Arbitrary, object-encoded thing's data.
        status:
          type: string
          description: Thing Status
          format: string
          example: enabled
        created_at:
          type: string
          format: date-time
          example: "2019-11-26 13:31:52"
          description: Time when the channel was created.
        updated_at:
          type: string
          format: date-time
          example: "2019-11-26 13:31:52"
          description: Time when the channel was created.
      xml:
        name: thing
...
    Channel:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: bb7edb32-2eac-4aad-aebe-ed96fe073879
          description: Unique channel identifier generated by the service.
        name:
          type: string
          example: channelName
          description: Free-form channel name. Channel name is unique on the given hierarchy level.
        domain_id:
          type: string
          format: uuid
          example: bb7edb32-2eac-4aad-aebe-ed96fe073879
          description: ID of the domain to which the group belongs.
        parent_id:
          type: string
          format: uuid
          example: bb7edb32-2eac-4aad-aebe-ed96fe073879
          description: Channel parent identifier.
        description:
          type: string
          example: long channel description
          description: Channel description, free form text.
        metadata:
          type: object
          example: { "role": "general" }
          description: Arbitrary, object-encoded channels's data.
        path:
          type: string
          example: bb7edb32-2eac-4aad-aebe-ed96fe073879.bb7edb32-2eac-4aad-aebe-ed96fe073879
          description: Hierarchy path, concatenated ids of channel ancestors.
        level:
          type: integer
          description: Level in hierarchy, distance from the root channel.
          format: int32
          example: 2
          maximum: 5
        created_at:
          type: string
          format: date-time
          example: "2019-11-26 13:31:52"
          description: Datetime when the channel was created.
        updated_at:
          type: string
          format: date-time
          example: "2019-11-26 13:31:52"
          description: Datetime when the channel was created.
        status:
          type: string
          description: Channel Status
          format: string
          example: enabled
      xml:
        name: channel
...
    ThingUpdate:
      type: object
      properties:
        name:
          type: string
          example: thingName
          description: Thing name.
        metadata:
          type: object
          example: { "role": "general" }
          description: Arbitrary, object-encoded thing's data.
      required:
        - name
        - metadata
...
    ThingSecret:
      type: object
      properties:
        secret:
          type: string
          example: bb7edb32-2eac-4aad-aebe-ed96fe073879
          description: New thing secret.
      required:
        - secret
    ChannelUpdate:
      type: object
      properties:
        name:
          type: string
          example: channelName
          description: Free-form channel name. Channel name is unique on the given hierarchy level.
        description:
          type: string
          example: long description but not too long
          description: Channel description, free form text.
        metadata:
          type: object
          example: { "role": "general" }
          description: Arbitrary, object-encoded channels's data.
      required:
        - name
        - metadata
        - description
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      example: { "error": "malformed entity specification" }
...
  parameters:
    ThingID:
      name: thingID
      description: Unique thing identifier.
      in: path
      schema:
        type: string
        format: uuid
        minLength: 36
        maxLength: 36
        pattern: "^[a-f0-9]{8}-[a-f0-9]{4}-[1-5][a-f0-9]{3}-[89ab][a-f0-9]{3}-[a-f0-9]{12}$"
      required: true
      example: bb7edb32-2eac-4aad-aebe-ed96fe073879
...
    chanID:
      name: chanID
      description: Unique channel identifier.
      in: path
      schema:
        type: string
        format: uuid
      required: true
      example: bb7edb32-2eac-4aad-aebe-ed96fe073879
...
  requestBodies:
...
    ThingUpdateReq:
      description: JSON-formated document describing the metadata and name of thing to be update
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ThingUpdate"
...
    ThingUpdateSecretReq:
      description: Secret change data. Thing can change its secret.
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ThingSecret"
...
    ChannelUpdateReq:
      description: JSON-formated document describing the metadata and name of channel to be update
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ChannelUpdate"
...
  responses:
...
    ThingRes:
      description: Data retrieved.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Thing"
      links:
        get_channels:
          operationId: listChannelsConnectedToThing
          parameters:
            thingID: $response.body#/id
        share:
          operationId: shareThing
          parameters:
            thingID: $response.body#/id
        unsahre:
          operationId: unshareThing
          parameters:
            thingID: $response.body#/id
...
    ChannelRes:
      description: Data retrieved.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Channel"
      links:
        get_things:
          operationId: listThingsInaChannel
          parameters:
            chanID: $response.body#/id
        get_users:
          operationId: listChannelsConnectedToUser
          parameters:
            memberID: $response.body#/id
        get_groups:
          operationId: listChannelsConnectedToGroup
          parameters:
            memberID: $response.body#/id
        assign_users:
          operationId: assignUsersToChannel
          parameters:
            chanID: $response.body#/id
        unassign_users:
          operationId: unassignUsersFromChannel
          parameters:
            chanID: $response.body#/id
        assign_groups:
          operationId: assignGroupsToChannel
          parameters:
            chanID: $response.body#/id
        unassign_groups:
          operationId: unassignGroupsFromChannel
          parameters:
            chanID: $response.body#/id
...
    ServiceError:
      description: Unexpected server-side error occurred.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        * Thing access: "Authorization: Bearer <user_access_token>"
security:
  - bearerAuth: []